requirement.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
numpy==1.17.4
Pillow==6.2.1
requests==2.22.0
torch==1.3.1
tqdm==4.40.0
urllib3==1.25.7
wget==3.2
